+++ src/main/java/frc/robot/subsystems/Drivetrain.java
@@ -26,8 +26,6 @@ import frc.robot.util.TestType;
 */
 public class Drivetrain extends SubsystemBase {
 
-  public boolean isSlewDrive = false; //TODO: what is this variable used for? It was here for testing slew but that isn't needed anymore
-
   // Swerve modules and other
   public SwerveModuleState[] m_swerveModuleStates = new SwerveModuleState[] {
     new SwerveModuleState(),
@@ -87,7 +85,7 @@ public class Drivetrain extends SubsystemBase {
 
   // Characterizing
   private boolean m_isCharacterizing = false;
-  private double m_charactericationVolts = 0; //TODO: read this variable name, slowly.
+  private double m_characterizationVolts = 0;
 
   public Drivetrain() {
     m_odometry = new SwerveDriveOdometry(m_kinematics, m_pigeon.getRotation2d(), getModulePositions());
@@ -98,7 +96,7 @@ public class Drivetrain extends SubsystemBase {
 
     if (m_isCharacterizing) {
       for (Module module : m_modules) {
-        module.characterize(m_charactericationVolts);
+        module.characterize(m_characterizationVolts);
       }
     } else {
       updateOdometry();
@@ -217,13 +215,6 @@ public class Drivetrain extends SubsystemBase {
   }
 
   /**
-   * Toggles the drive mode.
-   */
-  public void toggleDriveMode() {
-    isSlewDrive = !isSlewDrive;
-  }
-
-  /**
    * Gets the current robot pose from the odometry.
    */
   public Pose2d getPose() {
@@ -262,10 +253,13 @@ public class Drivetrain extends SubsystemBase {
    * @return an array of all swerve module positions
    */
   public SwerveModulePosition[] getModulePositions() {
-    SwerveModulePosition[] positions = new SwerveModulePosition[4]; // TODO: use {} to make this more concise
-    for (int i = 0; i < 4; i++) {
-      positions[i] = new SwerveModulePosition(); // TODO: this is just making an array of empty positions
-    }
+    // TODO: Need to test if this works
+    SwerveModulePosition[] positions = new SwerveModulePosition[]{
+      new SwerveModulePosition(m_modules[0].getDriveVelocity(), Rotation2d.fromDegrees(m_modules[0].getAngle())),
+      new SwerveModulePosition(m_modules[1].getDriveVelocity(), Rotation2d.fromDegrees(m_modules[1].getAngle())),
+      new SwerveModulePosition(m_modules[2].getDriveVelocity(), Rotation2d.fromDegrees(m_modules[2].getAngle())),
+      new SwerveModulePosition(m_modules[3].getDriveVelocity(), Rotation2d.fromDegrees(m_modules[3].getAngle()))
+    };
     return positions;
   }
 
@@ -293,7 +287,7 @@ public class Drivetrain extends SubsystemBase {
   /** Runs forwards at the commanded voltage. */
   public void runCharacterizationVolts(double volts) {
     m_isCharacterizing = true;
-    m_charactericationVolts = volts;
+    m_characterizationVolts = volts;
   }
 
   /** Returns the average drive velocity in radians/sec. */
